Move types to shared component
