+++ web/app/components/document/sidebar.hbs
@@ -299,7 +299,6 @@
           @documentIsDraft={{this.isDraft}}
           @productArea={{@document.product}}
           @objectID={{@document.objectID}}
-          @scope="all"
           @headerTitle="Related resources"
           @modalHeaderTitle="Add related resource"
           @modalInputPlaceholder="Search docs or paste a URL..."
