+++ web/app/components/document/sidebar/related-resources.hbs
@@ -6,7 +6,6 @@
   @modalInputPlaceholder={{@modalInputPlaceholder}}
   @addResource={{this.addResource}}
   @documentObjectID={{@objectID}}
-  @scope={{@scope}}
 >
   <:header as |rr|>
     <Document::Sidebar::SectionHeader
